+++ pages/mypage/index.tsx
@@ -4,11 +4,10 @@ import Header from '@/components/common/Header/SecondHeader/SecondHeader';
 import SideMenu from '@/components/common/SideMenu/SideMenu';
 import ProfileCard from '@/components/pages/mypage/ProfileCard';
 import PasswordCard from '@/components/pages/mypage/PasswordCard';
+import BackButton from '@/components/pages/mypage/BackButton';
 import { getUserInfo } from '@/api/users/getUserInfo';
 import { UserType } from '@/lib/types/users';
 import { DEVICE_SIZE } from '@/styles/DeviceSize';
-import { FONT_16 } from '@/styles/FontStyles';
-import BackwordIcon from '@/public/icon/arrow_backward.svg';
 import dashboardData from '@/components/common/SideMenu/mock';
 
 function MyPage() {
@@ -42,18 +41,6 @@ function MyPage() {
 
 export default MyPage;
 
-/**
- * TODO: 컴포넌트로 따로 빼고 & 애니메이션 & 링크 설정
- */
-function BackButton() {
-  return (
-    <StyledBackWrapper>
-      <BackwordIcon />
-      <StyledBackText>돌아가기</StyledBackText>
-    </StyledBackWrapper>
-  );
-}
-
 const StyledBody = styled.div`
   padding-top: 70px;
   padding-left: 300px;
@@ -77,16 +64,6 @@ const StyledContainer = styled.div`
   @media (max-width: ${DEVICE_SIZE.mobile}) {
     gap: 20px;
   }
-`;
-
-const StyledBackWrapper = styled.div`
-  display: flex;
-  align-items: center;
-  gap: 6px;
-`;
-
-const StyledBackText = styled.div`
-  ${FONT_16};
 `;
 
 const StyledCardWrapper = styled.div`
